% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIndices.R
\name{appendAttrIndexRecord}
\alias{appendAttrIndexRecord}
\title{Append a attr_index record to the attrIndexTable object}
\usage{
appendAttrIndexRecord(
  attrIndexTable,
  geomIndexTable,
  dataSourceRecord,
  variableSourceRecord
)
}
\arguments{
\item{attrIndexTable}{(data.frame) the existing attrIndexTable}

\item{geomIndexTable}{(data.frame) the existing geomIndexTable}

\item{dataSourceRecord}{(data.frame) A full record (entire row) from the backbone.data_source table}
}
\value{
(data.frame) updated attrIndexTable
}
\description{
Append a attr_index record to the attrIndexTable object
}
