% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIndices.R
\name{createAttrIndexRecord}
\alias{createAttrIndexRecord}
\title{Create a single record in the backbone.attr_index table}
\usage{
createAttrIndexRecord(conn, rec)
}
\arguments{
\item{conn}{(DatabaseConnectorJdbcConnection) A database connection object created with \code{DatabaseConnector::connect} function}

\item{rec}{(data.frame) A full record (entire row) from the backbone.data_source table}
}
\value{
A new record in the backbone.attr_index table
}
\description{
Create a single record in the backbone.attr_index table
}
\examples{
\dontrun{

record <- getDataSourceRecord(conn = conn, dataSourceUuid = 9999)

createAttrIndexRecord(conn = conn, rec = record)
}

}
