% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{createAttrInstanceTable}
\alias{createAttrInstanceTable}
\title{Create an empty instance of the attr_template in a given schema in PostGIS}
\usage{
createAttrInstanceTable(connectionDetails, schema, name)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{schema}{(character) The name of the schema in PostGIS where the empty attr_X should be created}

\item{name}{(character) The suffix for the table name to be created; will be appended to "attr_"}
}
\value{
An empty table (attr_X) in PostGIS
}
\description{
Create an empty instance of the attr_template in a given schema in PostGIS
}
\examples{
\dontrun{

variableSourceId <- 157

variableTable <- DatabaseConnector::dbGetQuery(conn, paste0("SELECT * FROM backbone.variable_source WHERE variable_source_id = ", variableSourceId))

attrIndex <- DatabaseConnector::dbGetQuery(conn, paste0("SELECT * FROM backbone.attr_index WHERE data_source_id = ", variableTable$data_source_uuid,";"))

createAttrInstanceTable(conn, schema = attrIndex$database_schema, name = attrIndex$table_name)
}

}
