% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExposureOccurrence.R, R/createGaiaDdl.R
\name{createOccurrencePrimaryKeys}
\alias{createOccurrencePrimaryKeys}
\alias{createOccurrenceForeignKeys}
\alias{createDdl}
\alias{createPrimaryKeys}
\alias{createForeignKeys}
\title{Create the OHDSI-SQL Common Data Model DDL code}
\usage{
createOccurrencePrimaryKeys(gaiaVersion)

createOccurrenceForeignKeys(gaiaVersion)

createDdl(gaiaVersion)

createPrimaryKeys(gaiaVersion)

createForeignKeys(gaiaVersion)
}
\arguments{
\item{gaiaVersion}{The version of gaiaDB you are creating, e.g. 001}
}
\value{
A character string containing postgresql DDL
}
\description{
The createDdl, createForeignKeys, and createPrimaryKeys functions each return a character string
containing their respective DDL SQL code in OHDSQL dialect for a specific CDM version.
The SQL they generate needs to be rendered and translated before it can be executed.
}
\details{
The DDL SQL code is created from a two csv files that detail the OMOP CDM Specifications.
These files also form the basis of the CDM documentation and the Data Quality
Dashboard.
}
\section{Functions}{
\itemize{
\item \code{createOccurrencePrimaryKeys()}: createPrimaryKeys Returns a string containing the OHDSQL for creation of primary keys in the OMOP CDM.

\item \code{createOccurrenceForeignKeys()}: createForeignKeys Returns a string containing the OHDSQL for creation of foreign keys in the OMOP CDM.

\item \code{createPrimaryKeys()}: createPrimaryKeys Returns a string containing the OHDSQL for creation of primary keys in the OMOP CDM.

\item \code{createForeignKeys()}: createForeignKeys Returns a string containing the OHDSQL for creation of foreign keys in the OMOP CDM.

}}
\examples{
ddl <- createDdl("001")
pk <- createPrimaryKeys("001")
fk <- createForeignKeys("001")
}
