% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{createGeomTableIndexes}
\alias{createGeomTableIndexes}
\title{Create indexes on a geom_* table}
\usage{
createGeomTableIndexes(connectionDetails, schema, name)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{schema}{(character) The name of the schema}

\item{name}{(character) The geometry table name}
}
\value{
Indexes on two geometry columns (geom_local_value, geom_wgs84)
}
\description{
Create indexes on a geom_* table
}
