% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIndices.R
\name{createIndices}
\alias{createIndices}
\title{Create and populate backbone.geom_index and backbone.attr_index from registered sources in backbone.data_source}
\usage{
createIndices(connectionDetails, uuids)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{uuids}{(list) A list of UUIDs for the data sources that are registered in the backbone.data_source table}
}
\value{
Two tables (backbone.geom_index and backbone.attr_index) in PostGIS
}
\description{
Create and populate backbone.geom_index and backbone.attr_index from registered sources in backbone.data_source
}
\examples{
\dontrun{
createIndices(connectionDetails = connectionDetails, uuids = get_uuids())
}

}
