% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createNameString}
\alias{createNameString}
\title{Sanitize and create table name from source values}
\usage{
createNameString(name)
}
\arguments{
\item{name}{(character) Table name source value}
}
\value{
(character) Sanitized table name string
}
\description{
Sanitize and create table name from source values
}
