% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageData.R
\name{createSpecTable}
\alias{createSpecTable}
\title{Reformat the attr_ or geom_spec JSON as a table}
\usage{
createSpecTable(jsonStringSpec)
}
\arguments{
\item{jsonStringSpec}{(JSON) The attr_ or geom_spec from the backbone.variable_source or backbone.data_source table.}
}
\value{
(data.frame) Table representation of the JSON attr or geom spec that defines how the source data can be programmatically standardized
}
\description{
Reformat the attr_ or geom_spec JSON as a table
}
\examples{

\dontrun{
specTable <- createSpecTable(dataSourceRecord$geom_spec)
}

}
