% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIndices.R
\name{get_uuids}
\alias{get_uuids}
\title{Get all UUIDs from backbone.data_source}
\usage{
get_uuids()
}
\value{
(list) All UUIDs from backbone.data_source
}
\description{
Get all UUIDs from backbone.data_source
}
