% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getAttrIndex}
\alias{getAttrIndex}
\title{Get the backbone.attr_index table}
\usage{
getAttrIndex(connectionDetails)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
(data.frame) The entire backbone.attr_index table
}
\description{
Get the backbone.attr_index table
}
