% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getAttrIndexRecord}
\alias{getAttrIndexRecord}
\title{Get a single record from the backbone.attr_index table}
\usage{
getAttrIndexRecord(connectionDetails, variableSourceId)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{dataSourceUuid}{(UUID) The UUID for the data source that is registered in the backbone.attr_index table}
}
\value{
(data.frame) A full record (entire row) from the backbone.attr_index table
}
\description{
Get a single record from the backbone.attr_index table
}
