% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getAttrNameFromVariableSourceId}
\alias{getAttrNameFromVariableSourceId}
\title{Get the name of a variable_source's attr dependency}
\usage{
getAttrNameFromVariableSourceId(connectionDetails, variableSourceId)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
(character) The schema and table name for the source attr dependency
}
\description{
Get the name of a variable_source's attr dependency
}
\examples{
\dontrun{
attrName <- getAttrNameFromVariableSourceId(connectionDetails, 101)
}

}
