% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getAttrOfGeomIndexId}
\alias{getAttrOfGeomIndexId}
\title{Get foreign key for attr_of_geom_index_id}
\usage{
getAttrOfGeomIndexId(connectionDetails, dataSourceUuid)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{dataSourceUuid}{(UUID) The UUID for the data source that is registered in the backbone.data_source table}
}
\value{
(integer) Identifier for the corresponding backbone.geom_index entry
}
\description{
Get foreign key for attr_of_geom_index_id
}
