% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getAttrTemplate}
\alias{getAttrTemplate}
\title{Get the attr_template table}
\usage{
getAttrTemplate(conn)
}
\arguments{
\item{conn}{(DatabaseConnectorJdbcConnection) A database connection object created with \code{DatabaseConnector::connect} function}
}
\value{
(data.frame) An empty attr_template table
}
\description{
Get the attr_template table
}
