% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getDataSourceRecord}
\alias{getDataSourceRecord}
\title{Get a single record from the backbone.data_source table}
\usage{
getDataSourceRecord(conn, dataSourceUuid)
}
\arguments{
\item{conn}{(DatabaseConnectorJdbcConnection) A database connection object created with \code{DatabaseConnector::connect} function}

\item{dataSourceUuid}{(UUID) The UUID for the data source that is registered in the backbone.data_source table}
}
\value{
(data.frame) A full record (entire row) from the backbone.data_source table
}
\description{
Get a single record from the backbone.data_source table
}
