% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getForeignKeyGid}
\alias{getForeignKeyGid}
\title{Get foreign key for attr_of_geom_index_id}
\usage{
getForeignKeyGid(conn, uuid)
}
\arguments{
\item{conn}{(DatabaseConnectorJdbcConnection) A database connection object created with \code{DatabaseConnector::connect} function}

\item{uuid}{(UUID) The UUID for the data source that is registered in the backbone.data_source table}
}
\value{
(integer) Identifier for the corresponding backbone.geom_index entry
}
\description{
Get foreign key for attr_of_geom_index_id
}
