% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getGeomIdMap}
\alias{getGeomIdMap}
\title{Create a mapping table between identifiers and values for a given geom_X}
\usage{
getGeomIdMap(connectionDetails, geomIndexId)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{geomIndexId}{(integer) Identifier of a record in the backbone.geom_index table. Usually sourced from the \code{attr_of_geom_index_id} entry of an attr_index record}
}
\value{
(data.frame) A mapping table between geom_record_id for a given geom_X and that record's source value
}
\description{
Create a mapping table between identifiers and values for a given geom_X
}
\examples{
\dontrun{
geomIdMap <- getGeomIdMap(conn = conn, geomIndex = 1)
}

}
