% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getGeomIndex}
\alias{getGeomIndex}
\title{Get the backbone.geom_index table}
\usage{
getGeomIndex(connectionDetails)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
(data.frame) The entire backbone.geom_index table
}
\description{
Get the backbone.geom_index table
}
