% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getGeomIndexByDataSourceUuid}
\alias{getGeomIndexByDataSourceUuid}
\title{Get a record from the backbone.geom_index table from it's corresponding UUID in backbone.data_source table}
\usage{
getGeomIndexByDataSourceUuid(conn, dataSourceUuid)
}
\arguments{
\item{conn}{(DatabaseConnectorJdbcConnection) A database connection object created with \code{DatabaseConnector::connect} function}

\item{dataSourceUuid}{(UUID) The UUID for the data source that is registered in the backbone.data_source table}
}
\value{
(data.frame) A full record (entire row) from the backbone.geom_index table
}
\description{
Get a record from the backbone.geom_index table from it's corresponding UUID in backbone.data_source table
}
