% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getGeomIndexTable}
\alias{getGeomIndexTable}
\title{Get the backbone.geom_index table}
\usage{
getGeomIndexTable(connectionDetails)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
(data.frame) The entire backbone.geom_index table
}
\description{
Get the backbone.geom_index table
}
