% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getGeomNameFromVariableSourceId}
\alias{getGeomNameFromVariableSourceId}
\title{Get the name of a variable_source's geom dependency}
\usage{
getGeomNameFromVariableSourceId(connectionDetails, variableSourceId)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{variableSourceId}{(integer) Identifier of a record in the variable_source table}
}
\value{
(character) The schema and table name for the source geom dependency
}
\description{
Get the name of a variable_source's geom dependency
}
\examples{
\dontrun{
geomName <- getGeomNameFromVariableSourceId(connectionDetails, 101)
}

}
