% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getGeomTemplate}
\alias{getGeomTemplate}
\title{Get the geom_template table}
\usage{
getGeomTemplate(conn)
}
\arguments{
\item{conn}{(DatabaseConnectorJdbcConnection) A database connection object created with \code{DatabaseConnector::connect} function}
}
\value{
(data.frame) An empty geom_template table
}
\description{
Get the geom_template table
}
