% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getUniqueVariablesInAttrX}
\alias{getUniqueVariablesInAttrX}
\title{Get list of unique variable IDs in attr_X}
\usage{
getUniqueVariablesInAttrX(connectionDetails, databaseSchema, tableName)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{databaseSchema}{(character) schema that contains an attr_X table}

\item{tableName}{(character) name of an attr_X table}
}
\value{
(vector) a character vector of all loaded variable source IDs
}
\description{
Get list of unique variable IDs in attr_X
}
