% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getVariableSource}
\alias{getVariableSource}
\title{Get the backbone.variable_source table}
\usage{
getVariableSource(connectionDetails)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
(data.frame) The entire backbone.variable_source table
}
\description{
Get the backbone.variable_source table
}
