% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getVariableSourceRecord}
\alias{getVariableSourceRecord}
\title{Get a single record from the backbone.variable_source table}
\usage{
getVariableSourceRecord(connectionDetails, variableSourceId)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{variableSourceId}{(integer) The identifier for the data source that is registered in the backbone.variable_source table}
}
\value{
(data.frame) A full record (entire row) from the backbone.variable_source table
}
\description{
Get a single record from the backbone.variable_source table
}
