% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getVariableSourceSummaryTable}
\alias{getVariableSourceSummaryTable}
\title{Get a summary table of select variable source records by ID}
\usage{
getVariableSourceSummaryTable(connectionDetails, variableSourceIds)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{loadedVariables}{(vector) a character vector of all loaded variable source IDs}
}
\value{
(data.frame) A table of select variable source records
}
\description{
Get a summary table of select variable source records by ID
}
