% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{handleShapefileImportJob}
\alias{handleShapefileImportJob}
\title{Handle the joining of geometry and attribute, as well as the simple or iterative import of a shapefile}
\usage{
handleShapefileImportJob(
  connectionDetails,
  attrTableString,
  geomTableString,
  variable
)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{attrTableString}{(character) Name of the attr_X table in databaseSchema.tableName format}

\item{geomTableString}{(character) Name of the geom_X table in databaseSchema.tableName format}

\item{variable}{(character) Name of the variable to be imported}
}
\value{
(sf, data.frame) An sf object consisting of a single attribute and geometry; the result of joining attr_X and geom_X from the PostGIS
}
\description{
Handle the joining of geometry and attribute, as well as the simple or iterative import of a shapefile
}
