% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{importAttrIndexTable}
\alias{importAttrIndexTable}
\title{Import the attr_index table to PostGIS}
\usage{
importAttrIndexTable(connectionDetails, attrIndexTable)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{attrIndexTable}{(data.frame) The entire backbone.attr_index table}
}
\value{
A populated backbone.attr_index table in PostGIS
}
\description{
Import the attr_index table to PostGIS
}
