% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{importAttrTable}
\alias{importAttrTable}
\title{Import a well-formatted attribute table into an empty instance of attr_X in PostGIS}
\usage{
importAttrTable(connectionDetails, attribute, attrIndex)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{attribute}{(data.frame) A well-formatted attribute. Created by appending staging data table to an attr_template}

\item{attrIndex}{(data.frame) A full record (entire row) from the backbone.attr_index table corresponding to the registered attribute of interest}
}
\value{
A table (attr_X) in PostGIS
}
\description{
Import a well-formatted attribute table into an empty instance of attr_X in PostGIS
}
