% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{importCohortTable}
\alias{importCohortTable}
\title{Import a geocoded cohort table to gaiaDB}
\usage{
importCohortTable(gaiaConnectionDetails, cohort, overwrite = FALSE)
}
\arguments{
\item{gaiaConnectionDetails}{(list) For connecting to gaiaDB. An object of class
connectionDetails as created by the
createConnectionDetails function}

\item{cohort}{(data.frame) An OHDSI cohort table with a POINT geometry column
named geometry}

\item{overwrite}{(boolean) Overwrite existing tables in the cohort schema?}
}
\value{
A new cohort table in the gaiaDB cohort schema
}
\description{
Import a geocoded cohort table to gaiaDB
}
\examples{
\dontrun{
importCohortTable(gaiaConnectionDetails = gaiaConnectionDetails,
                  cohort = geocodedCohort,
                  overwrite = FALSE)
}
}
