% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importShapefile.R
\name{importShapefile}
\alias{importShapefile}
\title{Import an attribute-geometry pair from PostGIS}
\usage{
importShapefile(connectionDetails, variableSourceId)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{variableSourceId}{(integer) The identifier for the variable that is registered in the PostGIS variable_source table}
}
\value{
(sf, data.frame) An sf object consisting of a single attribute and geometry; the result of joining attr_X and geom_X from the PostGIS
}
\description{
Import an attribute-geometry pair from PostGIS
}
\examples{

\dontrun{
shape <- importShapefile(connectionDetails = connectionDetails, variableSourceId = 157)
}

}
