% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{initializeDatabase}
\alias{initializeDatabase}
\title{Initialize gaiaDB}
\usage{
initializeDatabase(connectionDetails, overwrite = FALSE, testing = FALSE)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{overwrite}{(boolean) Potential loss of data, proceed with caution! Drop the backbone schema and reinitialize? You will lose any custom data or variable sources that you have added if you overwrite the backbone schema.}
}
\value{
(database schema) The gaiaDB backbone schema is added to the database in connectionDetails
}
\description{
Initialize gaiaDB
}
\examples{
\dontrun{
 connectionDetails <- DatabaseConnector::createConnectionDetails(
   dbms = "postgresql",
   server = "", # name of the server
   port = 5432,
   user="postgres", # username to access server
   password = "mysecretpassword")

 initializeDatabase(connectionDetails)
}

}
