% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{insertPostgisGeometry}
\alias{insertPostgisGeometry}
\title{Connect to the Postgis database and insert a geometry table}
\usage{
insertPostgisGeometry(connectionDetails, staged, geomIndex)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{staged}{(data.frame) A well-formatted geometry table. Created by appending staging data table to an geom_template}

\item{geomIndex}{(data.frame) A full record (entire row) from the backbone.geom_index table corresponding to the registered geometry of interest}
}
\value{
A table (geom_X) in PostGIS
}
\description{
Connect to the Postgis database and insert a geometry table
}
