% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSupportedVersions.R
\name{listSupportedDialects}
\alias{listSupportedDialects}
\title{List SQL dialects supported by this package}
\usage{
listSupportedDialects()
}
\value{
A character vector of supported SQL dialects.
}
\description{
Returns the SQL dialects supported via SqlRender, which this package can use to generate dialect-specific DDL.
}
