% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSupportedVersions.R
\name{listSupportedVersions}
\alias{listSupportedVersions}
\title{List OMOP CDM versions supported by this package}
\usage{
listSupportedVersions()
}
\value{
A character vector of supported CDM versions (e.g., "5.3", "5.4").
}
\description{
Returns the list of OMOP Common Data Model versions for which the package provides DDL and constraint support.
}
