% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVariable.R
\name{loadVariable}
\alias{loadVariable}
\title{Loads a variable from a registered source into PostGIS as attr_X}
\usage{
loadVariable(connectionDetails, variableSourceId)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}

\item{variableSourceId}{(integer) The identifier for the variable that is registered in the PostGIS variable_source table}
}
\value{
A table (attr_X) in PostGIS
}
\description{
Loads a variable from a registered source into PostGIS as attr_X
}
\examples{
\dontrun{

variableSourceId <- 157

variableSourceRecord <- DatabaseConnector::dbGetQuery(conn, paste0("SELECT * FROM backbone.variable_source WHERE variable_source_id = ", variableSourceId))

attrIndex <- DatabaseConnector::dbGetQuery(conn, paste0("SELECT * FROM backbone.attr_index WHERE data_source_id = ", variableSourceRecord$data_source_uuid,";"))

tableExists <- DatabaseConnector::existsTable(conn,
                                              attrIndex$database_schema,
                                              paste0("attr_", attrIndex$table_name))

if (!tableExists) {
  message("Loading attr table dependency")
  loadVariable(connectionDetails = connectionDetails, variableSourceId)
}
}

}
