% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageData.R
\name{readFromZip}
\alias{readFromZip}
\title{Unzip and read contents of a zipfile into R memory}
\usage{
readFromZip(zipfile, exdir, rec)
}
\arguments{
\item{zipfile}{(character) path to the compressed file}

\item{exdir}{(character) path to where contents of the compressed file should be extracted}

\item{rec}{(data.frame) A full record (entire row) from the backbone.data_source table corresponding to the data source of interest. Usually created using \code{getDataSourceRecord} function}
}
\value{
(data.frame) An untransformed version of the source data
}
\description{
Unzip and read contents of a zipfile into R memory
}
