% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageData.R
\name{readStorageConfig}
\alias{readStorageConfig}
\title{Read config.yml file from inst}
\usage{
readStorageConfig()
}
\value{
(list) User-defined storage configuration
}
\description{
Read config.yml file from inst
}
