% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageData.R
\name{standardizeStaged}
\alias{standardizeStaged}
\title{Create a standardized version of the staged data that was imported from source}
\usage{
standardizeStaged(staged, specTable)
}
\arguments{
\item{staged}{(data.frame) Table of either attributes or geometries that was imported from a source by \code{getStaged} function}

\item{specTable}{(data.frame) Table representation of the JSON attr or geom spec that defines how the source data can be programmatically standardized}
}
\value{
(data.frame) A table standardized in the attr_ or geom_template mold
}
\description{
Create a standardized version of the staged data that was imported from source
}
\examples{

\dontrun{
stagedResult <- standardizeStaged(staged = staged, specTable = specTable)
}

}
