% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageData.R
\name{standardizeStagedGeom}
\alias{standardizeStagedGeom}
\title{Create a standardized version of the staged data that was imported from source}
\usage{
standardizeStagedGeom(staged, geomSpec)
}
\arguments{
\item{staged}{(data.frame) Table of either attributes or geometries that was imported from a source by \code{getStaged} function}

\item{spec}{(JSON) The geom_spec from the backbone.data_source table.}
}
\value{
(data.frame) A table standardized in the geom_template mold
}
\description{
Create a standardized version of the staged data that was imported from source
}
\examples{

\dontrun{
stagedResult <- standardizeStagedGeom(staged = staged, spec = dataSourceRecord$geom_spec)
}

}
