% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGaiaDdl.R
\name{writeDdl}
\alias{writeDdl}
\alias{writePrimaryKeys}
\alias{writeForeignKeys}
\title{Write gaiaDB DDL script}
\usage{
writeDdl(gaiaVersion, outputfolder)

writePrimaryKeys(gaiaVersion, outputfolder)

writeForeignKeys(gaiaVersion, outputfolder)
}
\arguments{
\item{gaiaVersion}{The version of the CDM you are creating, e.g. 5.3, 5.4}

\item{outputfolder}{The directory or folder where the SQL file should be saved.}
}
\description{
Write the DDL to a SQL file. The SQL will be rendered (parameters replaced) and translated to the target SQL
dialect. By default the @cdmDatabaseSchema parameter is kept in the SQL file and needs to be replaced before
execution.
}
\section{Functions}{
\itemize{
\item \code{writePrimaryKeys()}: writePrimaryKeys Write the SQL code that creates the primary keys to a file.

\item \code{writeForeignKeys()}: writeForeignKeys Write the SQL code that creates the foreign keys to a file.

}}
