% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIndices.R
\name{getAttrOfGeomIndexId}
\alias{getAttrOfGeomIndexId}
\title{Get foreign key for attr_of_geom_index_id}
\usage{
getAttrOfGeomIndexId(geomIndexTable, dataSourceUuid)
}
\arguments{
\item{geomIndexTable}{(data.frame) the existing geomIndexTable}

\item{dataSourceUuid}{(UUID) The UUID for the geometry data source that is registered in the backbone.data_source table}
}
\value{
(integer) Identifier for the corresponding backbone.geom_index entry
}
\description{
Get foreign key for attr_of_geom_index_id
}
