% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getAttrTemplate}
\alias{getAttrTemplate}
\title{Get the attr_template table}
\usage{
getAttrTemplate(connectionDetails)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
(data.frame) An empty attr_template table
}
\description{
Get the attr_template table
}
