% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUtils.R
\name{getVariableSourceSummaryTable}
\alias{getVariableSourceSummaryTable}
\title{Get the Variable Source Summary table}
\usage{
getVariableSourceSummaryTable(connectionDetails)
}
\arguments{
\item{connectionDetails}{(list) An object of class connectionDetails as created by the createConnectionDetails function}
}
\value{
(data.frame) A table of variable source records and a boolean column for if they are loaded.
}
\description{
Get the Variable Source Summary table
}
\examples{
\dontrun{
variableSourceSummaryTable <- getVariableSourceSummaryTable(connectionDetails)
}

}
