% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExposureOccurrence.R, R/writeGaiaDdl.R
\name{writeOccurrencePrimaryKeys}
\alias{writeOccurrencePrimaryKeys}
\alias{writeOccurrenceForeignKeys}
\alias{writeOccurrenceIndex}
\alias{writeDdl}
\alias{writePrimaryKeys}
\alias{writeForeignKeys}
\title{Write gaiaDB DDL script}
\usage{
writeOccurrencePrimaryKeys(
  targetDialect,
  gaiaVersion,
  outputfolder,
  cdmDatabaseSchema = "@cdmDatabaseSchema"
)

writeOccurrenceForeignKeys(
  targetDialect,
  gaiaVersion,
  outputfolder,
  cdmDatabaseSchema = "@cdmDatabaseSchema"
)

writeOccurrenceIndex(
  targetDialect,
  gaiaVersion,
  outputfolder,
  cdmDatabaseSchema = "@cdmDatabaseSchema"
)

writeDdl(gaiaVersion, outputfolder)

writePrimaryKeys(gaiaVersion, outputfolder)

writeForeignKeys(gaiaVersion, outputfolder)
}
\arguments{
\item{gaiaVersion}{The version of the CDM you are creating, e.g. 5.3, 5.4}

\item{outputfolder}{The directory or folder where the SQL file should be saved.}
}
\description{
Write the DDL to a SQL file. The SQL will be rendered (parameters replaced) and translated to the target SQL
dialect. By default the @cdmDatabaseSchema parameter is kept in the SQL file and needs to be replaced before
execution.
}
\section{Functions}{
\itemize{
\item \code{writeOccurrencePrimaryKeys()}: writePrimaryKeys Write the SQL code that creates the primary keys to a file.

\item \code{writeOccurrenceForeignKeys()}: writeForeignKeys Write the SQL code that creates the foreign keys to a file.

\item \code{writeOccurrenceIndex()}: writeIndex Write the rendered and translated sql that creates recommended indexes to a file.

\item \code{writePrimaryKeys()}: writePrimaryKeys Write the SQL code that creates the primary keys to a file.

\item \code{writeForeignKeys()}: writeForeignKeys Write the SQL code that creates the foreign keys to a file.

}}
